test: split PDF equation source sequence coverage
